+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/descriptor/ServerSettingsDescriptor.java
@@ -22,12 +22,14 @@ import hudson.util.ListBoxModel;
 import jenkins.model.Jenkins;
 import lombok.NonNull;
 import org.apache.commons.lang.StringUtils;
+import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.QueryParameter;
 
 import java.util.Collections;
 import java.util.UUID;
 
+@Symbol("serverSettings")
 @Extension
 public class ServerSettingsDescriptor extends Descriptor<ServerSettings> {
     public ServerSettingsDescriptor() {