+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/descriptor/TransferDescriptor.java
@@ -23,7 +23,7 @@ public class TransferDescriptor extends Descriptor<Transfer> {
     @Override
     @NonNull
     public String getDisplayName() {
-        return "TransferDescriptor";
+        return "transferDescriptor";
     }
 
     public FormValidation doCheckIncludes(@QueryParameter final String includes) {