+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/localconfig/ConfigCustom.java
@@ -18,7 +18,7 @@ public class ConfigCustom extends ConfigBase {
         this.serverSettings = serverSettings;
     }
 
-    @Symbol("ConfigCustom")
+    @Symbol("configCustom")
     @Extension
     public static class Descriptor extends ConfigBaseDescriptor {
         @Override