+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/Data.java
@@ -35,7 +35,7 @@ public class Data extends BaseReport {
         this.filter = filter;
     }
 
-    @Symbol("Data")
+    @Symbol("data")
     @Extension
     public static class DataDescriptor extends BaseReportDescriptor {
         @NonNull